pyproject.toml
[project]
name = "rag-search-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.57.0",
    "nltk==3.9.1",
    "numpy>=2.4.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
]
